+++ src/main/java/com/haojiyou/cnchar/common/ReplaceCharConfig.java
@@ -16,7 +16,14 @@ public class ReplaceCharConfig {
      * 缓存替换中英文字符配置
      */
     public static Map<String, String> cnCharMap = new HashMap<>();
+    /**
+     * 是否显示替换提示
+     */
     public static boolean showRepacedMsg = false;
+    /**
+     * 注释区域是否替换
+     */
+    public static boolean replaceInComment = false;
 
     static {
         reload();
@@ -28,6 +35,8 @@ public class ReplaceCharConfig {
         for (int i = 0; i < configString.length / 2; i++) {
             cnCharMap.put(configString[2 * i].trim(), configString[2 * i + 1].trim());
         }
-        showRepacedMsg = Boolean.parseBoolean(PropertiesComponent.getInstance().getValue(CnCharSettingComponent.REPLACED_MSG_SHOW_KEY, "false"));
+        showRepacedMsg = PropertiesComponent.getInstance().getBoolean(CnCharSettingComponent.REPLACED_MSG_SHOW_KEY, false);
+
+        replaceInComment = PropertiesComponent.getInstance().getBoolean(CnCharSettingComponent.REPLACED_IN_COMMENT_KEY, false);
     }
 }